feat(smart-meter): moved device config modal to SmartMetersPage
