+++ src/pages/details/SmartMeterDetailsPage.tsx
@@ -1,6 +1,6 @@
 import { useActivePage } from '@toolpad/core';
 import { SmartMeterDto } from '../../api/openAPI';
-import { useParams, useSearchParams } from 'react-router-dom';
+import { useLocation, useParams } from 'react-router-dom';
 import { useSmartMeterService } from '../../hooks/services/useSmartMeterService.ts';
 import { useEffect, useMemo, useState } from 'react';
 import { useSnackbar } from '../../hooks/useSnackbar.ts';
@@ -13,7 +13,7 @@ const SmartMeterDetailsPage = () => {
     const [openAddMetadata, setOpenAddMetadata] = useState<boolean>(false);
 
     const params = useParams<{ id: string }>();
-    const searchParams = useSearchParams();
+    const location = useLocation();
     const activePage = useActivePage();
     const { showSnackbar } = useSnackbar();
     const { getSmartMeter } = useSmartMeterService();
@@ -39,10 +39,8 @@ const SmartMeterDetailsPage = () => {
         };
         void loadSmartMeter();
 
-        const open = searchParams[0].get('open');
-        if (open) {
+        if (location.state?.openDialog) {
             setOpenAddMetadata(true);
-            searchParams[0].delete('open');
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [params.id]);
